refactor: improve code readability and structure
