+++ one_fm/hooks.py
@@ -165,7 +165,7 @@ website_generators = ["Client"]
 # ------------
 
 before_install = "one_fm.install.before_install.execute"
-after_install = "one_fm.setup.after_install"
+after_install = "one_fm.setup.setup.after_install"
 #revert staging
 
 # Desk Notifications
