+++ one_fm/setup/assignment_rule.py
@@ -1,0 +1,41 @@
+from one_fm.custom.assignment_rule.assignment_rule import (
+	get_assignment_rule_json_file, create_assignment_rule, delete_assignment_rule
+)
+
+def create_assignment_rules():
+	create_assignment_rule(get_assignment_rule_json_file("roster_post_action_site_supervisor.json"))
+	create_assignment_rule(get_assignment_rule_json_file("subcontract_staff_shortlist.json"))
+	create_assignment_rule(get_assignment_rule_json_file("action_poc_check.json"))
+	create_assignment_rule(get_assignment_rule_json_file("shift_permission_approver.json"))
+	create_assignment_rule(get_assignment_rule_json_file("confirm_erpnext_task.json"))
+	create_assignment_rule(get_assignment_rule_json_file("leave_application_pending_hr.json"))
+	create_assignment_rule(get_assignment_rule_json_file("shift_request_pending_approval.json"))
+	create_assignment_rule(get_assignment_rule_json_file("attendance_request_approval.json"))
+	create_assignment_rule(get_assignment_rule_json_file("subcontract_staff_request.json"))
+	create_assignment_rule(get_assignment_rule_json_file("attendance_check_reports_to.json"))
+	create_assignment_rule(get_assignment_rule_json_file("attendance_check_site_supervisor.json"))
+	create_assignment_rule(get_assignment_rule_json_file("attendance_check_shift_supervisor.json"))
+	create_assignment_rule(get_assignment_rule_json_file("rfm_approver.json"))
+	create_assignment_rule(get_assignment_rule_json_file("purchase_order_purchase_manager_action.json"))
+	create_assignment_rule(get_assignment_rule_json_file("purchase_order_finance_manager_action.json"))
+	create_assignment_rule(get_assignment_rule_json_file("employee_checkin_issue_approval.json"))
+	create_assignment_rule(get_assignment_rule_json_file("shift_request_draft.json"))
+	create_assignment_rule(get_assignment_rule_json_file("default_shift_checker_site_supervisor.json"))
+	create_assignment_rule(get_assignment_rule_json_file("remove_blocker_assigned_to.json"))
+	create_assignment_rule(get_assignment_rule_json_file("leave_approver_assignment.json"))
+	create_assignment_rule(get_assignment_rule_json_file("career_history.json"))
+	create_assignment_rule(get_assignment_rule_json_file("leave_extension_request_approver.json"))
+	create_assignment_rule(get_assignment_rule_json_file("hd_ticket_process_owner.json"))
+	create_assignment_rule(get_assignment_rule_json_file("hd_ticket_bug_buster.json"))
+	create_assignment_rule(get_assignment_rule_json_file("post_schedule_checker_project_manager.json"))
+	create_assignment_rule(get_assignment_rule_json_file("roster_day_off_checker_project_manager.json"))
+	create_assignment_rule(get_assignment_rule_json_file("roster_day_off_checker_site_supervisor.json"))
+	create_assignment_rule(get_assignment_rule_json_file("roster_employee_actions_project_manager.json"))
+	create_assignment_rule(get_assignment_rule_json_file("roster_employee_actions_site_supervisor.json"))
+
+def delete_assignment_rules():
+	delete_assignment_rule(get_assignment_rule_json_file("roster_post_action_site_supervisor.json"))
+	delete_assignment_rule(get_assignment_rule_json_file("subcontract_staff_shortlist.json"))
+	delete_assignment_rule(get_assignment_rule_json_file("action_poc_check.json"))
+	delete_assignment_rule(get_assignment_rule_json_file("shift_permission_approver.json"))
+	delete_assignment_rule(get_assignment_rule_json_file("task.json"))
