+++ one_fm/setup/workflow.py
@@ -1,0 +1,64 @@
+from one_fm.custom.workflow.workflow import (
+	get_workflow_json_file, create_workflow, delete_workflow
+)
+
+def create_workflows():
+	create_workflow(get_workflow_json_file("erf.json"))
+	create_workflow(get_workflow_json_file("leave_acknowledgement_form.json"))
+	create_workflow(get_workflow_json_file("task.json"))
+	create_workflow(get_workflow_json_file("purchase_order.json"))
+	create_workflow(get_workflow_json_file("journal_entry.json"))
+	create_workflow(get_workflow_json_file("attendance_check.json"))
+	create_workflow(get_workflow_json_file("rfp.json"))
+	create_workflow(get_workflow_json_file("rfm.json"))
+	create_workflow(get_workflow_json_file("item.json"))
+	create_workflow(get_workflow_json_file("duty_commencement.json"))
+	create_workflow(get_workflow_json_file("request_employee_assignment.json"))
+	create_workflow(get_workflow_json_file("request_employee_schedule.json"))
+	create_workflow(get_workflow_json_file("mgrp.json"))
+	create_workflow(get_workflow_json_file("work_permit.json"))
+	create_workflow(get_workflow_json_file("medical_insurance.json"))
+	create_workflow(get_workflow_json_file("moi.json"))
+	create_workflow(get_workflow_json_file("paci.json"))
+	create_workflow(get_workflow_json_file("fingerprint_appointment.json"))
+	create_workflow(get_workflow_json_file("transfer_paper.json"))
+	create_workflow(get_workflow_json_file("pifss_form_103.json"))
+	create_workflow(get_workflow_json_file("employee_id.json"))
+	create_workflow(get_workflow_json_file("work_contract.json"))
+	create_workflow(get_workflow_json_file("job_offer.json"))
+	create_workflow(get_workflow_json_file("candidate_orientation.json"))
+	create_workflow(get_workflow_json_file("bank_account.json"))
+	create_workflow(get_workflow_json_file("missed_mom_follow_up_action_workflow.json"))
+	create_workflow(get_workflow_json_file("penalty.json"))
+	create_workflow(get_workflow_json_file("operations_site.json"))
+	create_workflow(get_workflow_json_file("shift_permission.json"))
+	create_workflow(get_workflow_json_file("gp_letter_request_wf.json"))
+	create_workflow(get_workflow_json_file("inovice.json"))
+	create_workflow(get_workflow_json_file("item_request_wf.json"))
+	create_workflow(get_workflow_json_file("purchase_request_wf.json"))
+	create_workflow(get_workflow_json_file("supplier_purchase_order_wf.json"))
+	create_workflow(get_workflow_json_file("legal_investigation.json"))
+	create_workflow(get_workflow_json_file("sales_invoice.json"))
+	create_workflow(get_workflow_json_file("asset.json"))
+	create_workflow(get_workflow_json_file("asset_movement.json"))
+	create_workflow(get_workflow_json_file("pifss_monthly_deduction.json"))
+	create_workflow(get_workflow_json_file("pifss_monthly_deduction_tool.json"))
+	create_workflow(get_workflow_json_file("overtime_request.json"))
+	create_workflow(get_workflow_json_file("employee_incentive.json"))
+	create_workflow(get_workflow_json_file("onboard_employee.json"))
+	create_workflow(get_workflow_json_file("shift_request.json"))
+	create_workflow(get_workflow_json_file("attendance_request.json"))
+	create_workflow(get_workflow_json_file("leave_application.json"))
+	create_workflow(get_workflow_json_file("timesheet.json"))
+	create_workflow(get_workflow_json_file("employee_checkin_issue.json"))
+	create_workflow(get_workflow_json_file("contracts.json"))
+	create_workflow(get_workflow_json_file("incident_form.json"))
+	create_workflow(get_workflow_json_file("subcontract_staff_shortlist.json"))
+	create_workflow(get_workflow_json_file("subcontract_staff_request.json"))
+	create_workflow(get_workflow_json_file("onboard_subcontract_employee.json"))
+	create_workflow(get_workflow_json_file("leave_extension_request.json"))
+
+def delete_workflows():
+	delete_workflow(get_workflow_json_file("erf.json"))
+	delete_workflow(get_workflow_json_file("leave_acknowledgement_form.json"))
+	delete_workflow(get_workflow_json_file("task.json"))
